Removing separate module, tidy
