+++ testsupport/src/main/java/org/sonatype/sisu/litmus/testsupport/concurrent/ConcurrentTestContext.java
@@ -10,7 +10,7 @@
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the Apache License Version 2.0 for the specific language governing permissions and limitations there under.
  */
-package org.sonatype.sisu.litmus.concurrenttest;
+package org.sonatype.sisu.litmus.testsupport.concurrent;
 
 import java.util.concurrent.BrokenBarrierException;
 import java.util.concurrent.CyclicBarrier;
@@ -23,6 +23,8 @@ import static com.google.common.base.Preconditions.checkNotNull;
 
 /**
  * Context shared between the {@link ConcurrentRunner} and the various {@link ConcurrentTestWorker}s.
+ *
+ * @since 1.15
  */
 class ConcurrentTestContext
 {
