+++ testsupport/src/test/java/org/sonatype/sisu/litmus/testsupport/concurrent/ConcurrentRunnerTest.java
@@ -10,11 +10,13 @@
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the Apache License Version 2.0 for the specific language governing permissions and limitations there under.
  */
-package org.sonatype.sisu.litmus.concurrenttest;
+package org.sonatype.sisu.litmus.testsupport.concurrent;
 
 import java.util.concurrent.atomic.AtomicInteger;
 
 import org.sonatype.sisu.litmus.testsupport.TestSupport;
+import org.sonatype.sisu.litmus.testsupport.concurrent.ConcurrentRunner;
+import org.sonatype.sisu.litmus.testsupport.concurrent.ConcurrentTask;
 
 import org.junit.Test;
